# search: cap groupsize workers by NumCPU / GOMAXPROCS